Handle filtering of connections correctly
Note! This probably doesn't work how we want (in particular with regards
to pagination). The main issue is that `scoped_items` is run before the
`after_resolve` hook of the connection field extension (which means it
has not been limited or paged yet) - further investigation is needed
but I strongly suspect that we will need to do this in a field
extension rather than in the scoped_items method.