Avoid the need to fake authorization in type spec
This was caused by the failure to pass the current user to the resolve
field mechanism. To fix this `GraphqlHelpers#resolve` and
`GraphqlHelpers#resolve_field` are now unified (resolve is defined in
terms of resolve_field), and we pass the current user explicitly in the
PrometheusIntegrationType spec.

An additional change is the deletion of the prometheus_service_policy,
which is not necessary given the fact that all Service instances now
have a suitable policy that delegates to project.