Global Search - Project Filter
This change replaces the deprecated
jQuery Dropdown plugin.

Instead we use gl-dropdown from
GitLab UI and a Vue component.

This uses a Vuex store to manage the
API calls and API data.  From there
the GitLab UI components take
care of the existing functionality.

The previous change focused on
the Group filter.
This change focueses on
the Project filter.