Responding to code review comments
Added dedicated `admin_user` instead of modifying existing user to be an
admin.

Added a `ctx:` parameter for `#resolve_users` so we can pass in an admin
user.

Check if the current user is an `admin` in `UsersFinder` so the Grpahql
and REST endpoints behave the same way.

Update `user_finder_shared_specs` to include an admin user. Updated
`users_finder_spec.rb` to include admin in results.

Updated graphql schema and docs.

Fixed failing spec with logic change in `UsersResolver`.