De-normalize project visibility_level in issues in Elasticsearch
This is necessary to search issues correctly without joining to the
parent project. The previous case where we only set
`issues_access_level` will not correctly handle "Internal" projects
because "Internal" does not change the `issues_access_level` of the
project and we need a way to ensure that issues from "Internal" projects
are only found by logged in users.